restructured and documented Net46 null objects and added debug flag to Net46 project so tests will pass
