+++ ReactWindows/ReactNative.Net46.Tests/ReactNative.Net46.Tests.csproj
@@ -124,6 +124,5 @@
     </ProjectReference>
   </ItemGroup>
   <Import Project="..\ReactNative.Shared.Tests\ReactNative.Shared.Tests.projitems" Label="Shared" />
-  <Import Project="..\ReactNative.Shared\ReactNative.Shared.projitems" Label="Shared" />
   <Import Project="$(MSBuildBinPath)\Microsoft.CSharp.targets" />
 </Project>
